Consolidate reward processing cron handlers into route
