+++ apps/web/app/(ee)/api/cron/rewards/process/route.ts
@@ -1,21 +1,23 @@
+import { notifyPartnerRewardChange } from "@/lib/api/partners/notify-partner-reward-change";
 import { rewardJobSchema } from "@/lib/api/rewards/queue-reward-processing.ts";
+import { CRON_BATCH_SIZE } from "@/lib/cron";
 import { withCron } from "@/lib/cron/with-cron";
+import { REWARD_EVENT_COLUMN_MAPPING } from "@/lib/zod/schemas/rewards";
 import { prisma } from "@dub/prisma";
+import { Prisma } from "@dub/prisma/client";
+import { sleep } from "@dub/utils";
 import { logAndRespond } from "../../utils";
-import { handleRewardCreated } from "./handle-reward-created.ts";
-import { handleRewardDeleted } from "./handle-reward-deleted.ts";
-import { handleRewardUpdated } from "./handle-reward-updated.ts";
 
 export const dynamic = "force-dynamic";
 
 export const maxDuration = 600;
 
-// POST /api/cron/rewards/sync-enrollments
+// POST /api/cron/rewards/process
 export const POST = withCron(async ({ rawBody }) => {
   const input = rewardJobSchema.parse(JSON.parse(rawBody));
 
   const { event, payload } = input;
-  const { groupId, rewardId } = payload;
+  const { groupId, rewardId, occurredAt, rewardSnapshot } = payload;
 
   const reward = await prisma.reward.findUnique({
     where: {
@@ -54,25 +56,114 @@ export const POST = withCron(async ({ rawBody }) => {
     return logAndRespond(`Group ${groupId} not found. Skipping...`);
   }
 
-  const params = {
-    payload,
-    program: group.program,
-    group,
-    reward,
-  };
+  const rewardIdColumn = REWARD_EVENT_COLUMN_MAPPING[reward.event];
+
+  let startingAfter: string | undefined = undefined;
+  let where: Prisma.ProgramEnrollmentWhereInput | undefined = undefined;
+  let data: Prisma.ProgramEnrollmentUpdateManyArgs["data"] | undefined =
+    undefined;
 
   switch (event) {
     case "reward-created":
-      await handleRewardCreated(params);
+      where = { [rewardIdColumn]: null };
+      data = { [rewardIdColumn]: reward.id };
       break;
 
     case "reward-updated":
-      await handleRewardUpdated(params);
+      where = { [rewardIdColumn]: reward.id };
       break;
 
     case "reward-deleted":
-      await handleRewardDeleted(params);
+      where = { [rewardIdColumn]: reward.id };
+      data = { [rewardIdColumn]: null };
       break;
+  }
+
+  while (true) {
+    const programEnrollments = await prisma.programEnrollment.findMany({
+      where: {
+        groupId: group.id,
+        ...where,
+      },
+      select: {
+        id: true,
+        partner: {
+          select: {
+            users: {
+              select: {
+                user: {
+                  select: {
+                    name: true,
+                    email: true,
+                  },
+                },
+              },
+            },
+          },
+        },
+      },
+      orderBy: {
+        id: "asc",
+      },
+      ...(startingAfter !== undefined && {
+        skip: 1,
+        cursor: {
+          id: startingAfter,
+        },
+      }),
+      take: CRON_BATCH_SIZE,
+    });
+
+    if (programEnrollments.length === 0) {
+      break;
+    }
+
+    // Only when event is "reward-created" or "reward-deleted"
+    if (data) {
+      await prisma.programEnrollment.updateMany({
+        where: {
+          id: {
+            in: programEnrollments.map(({ id }) => id),
+          },
+        },
+        data: {
+          ...data,
+        },
+      });
+    }
+
+    const users = programEnrollments.flatMap(({ partner }) =>
+      partner.users.map(({ user }) => user),
+    );
+
+    await notifyPartnerRewardChange({
+      action: event,
+      program: group.program,
+      reward,
+      rewardSnapshot,
+      effectiveAt: occurredAt,
+      users,
+    });
+
+    startingAfter = programEnrollments[programEnrollments.length - 1].id;
+
+    await sleep(500);
+  }
+
+  if (event === "reward-deleted") {
+    try {
+      await prisma.reward.delete({
+        where: {
+          id: reward.id,
+        },
+      });
+    } catch (error) {
+      if (error.code !== "P2025") {
+        throw new Error(
+          `Failed to hard delete reward ${reward.id}: ${error.message}`,
+        );
+      }
+    }
   }
 
   return logAndRespond(`Processed reward ${rewardId} for group ${groupId}.`);
